Dav helped me get the data out, I need to use the annotation @Inject, got the applications to show and the borrowers, employees next

+++ src/main/java/com/stackleader/loanapplication/Borrower.java
@@ -2,10 +2,8 @@ package com.stackleader.loanapplication;
 
 import java.util.ArrayList;
 import java.util.List;
-import javax.annotation.processing.Generated;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-// this import isn't working (Why?) import javax.xml.bind.annotation.XmlTransient; //TODO
 
 /*
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
@@ -21,7 +19,7 @@ public class Borrower {
     @Id
     @GeneratedValue
     //@XmlTransient // This won't work? Check line 8 // TODO
-    int id;
+    long id;
     String firstName;
     String lastName;
     int age;
@@ -96,11 +94,11 @@ public class Borrower {
     }
 
     //Trying to make this transent
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
     //Transient
